backend/requirements.txt
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
setuptools==75.1.0
#gunicorn==23.0.0
#psycopg2-binary==2.9.9
#whitenoise==6.7.0
